ref(wasm): split layout and input entrypoints
