+++ src/module-input.c
@@ -1,0 +1,6 @@
+/* module-input.c — input-only compilation unit, no Clay layout engine */
+
+#include "mem.c"
+#include "utf8.c"
+#include "trie.c"
+#include "input.c"
