+++ src/module-layout.c
@@ -1,0 +1,10 @@
+/* module-layout.c — layout-only compilation unit, no input parser */
+
+#include "../clay/clay.h"
+
+#include "mem.c"
+#include "buffer.c"
+#include "cell.c"
+#include "utf8.c"
+#include "wcwidth.c"
+#include "clayterm.c"
