+++ tasks/bundle-wasm.ts
@@ -1,6 +1,8 @@
 import { encodeBase64 } from "@std/encoding/base64";
 
-const wasm = await Deno.readFile("clayterm.wasm");
+const [input = "clayterm.wasm", output = "wasm.ts"] = Deno.args;
+
+const wasm = await Deno.readFile(input);
 const base64 = encodeBase64(wasm);
 
 const source = `const bin = atob("${base64}");
@@ -9,5 +11,5 @@ for (let i = 0; i < bin.length; i++) bytes[i] = bin.charCodeAt(i);
 export const compiled = await WebAssembly.compile(bytes);
 `;
 
-await Deno.writeTextFile("wasm.ts", source);
-console.log(`wrote wasm.ts (${wasm.length} bytes encoded)`);
+await Deno.writeTextFile(output, source);
+console.log(`wrote ${output} (${wasm.length} bytes encoded)`);
